Move bucket actions to a sync client
