+++ src/resources/bucket.rs
@@ -1,7 +1,5 @@
-use crate::error::{Error, GoogleResponse};
 use crate::resources::bucket_access_control::{BucketAccessControl, NewBucketAccessControl};
 pub use crate::resources::common::Entity;
-use crate::resources::common::ListResponse;
 use crate::resources::default_object_access_control::{
     DefaultObjectAccessControl, NewDefaultObjectAccessControl,
 };
@@ -12,7 +10,7 @@ pub use crate::resources::location::*;
 /// is a single global namespace shared by all buckets. For more information, see
 /// [Bucket Name Requirements](https://cloud.google.com/storage/docs/naming#requirements).
 ///
-/// Buckets contain objects which can be accessed by their own methods. In addition to the 
+/// Buckets contain objects which can be accessed by their own methods. In addition to the
 /// [ACL property](https://cloud.google.com/storage/docs/access-control/lists), buckets contain
 /// `BucketAccessControls`, for use in fine-grained manipulation of an existing bucket's access
 /// controls.
@@ -538,395 +536,4 @@ pub struct TestIamPermission {
     /// (https://cloud.google.com/storage/docs/access-control/iam-permissions) for a list of
     /// supported permissions.
     permissions: Vec<String>,
-}
-
-impl Bucket {
-    /// Creates a new `Bucket`. There are many options that you can provide for creating a new
-    /// bucket, so the `NewBucket` resource contains all of them. Note that `NewBucket` implements
-    /// `Default`, so you don't have to specify the fields you're not using. And error is returned
-    /// if that bucket name is already taken.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::bucket::{Bucket, NewBucket};
-    /// use cloud_storage::bucket::{Location, MultiRegion};
-    ///
-    /// let new_bucket = NewBucket {
-    ///    name: "cloud-storage-rs-doc-1".to_string(), // this is the only mandatory field
-    ///    location: Location::Multi(MultiRegion::Eu),
-    ///    ..Default::default()
-    /// };
-    /// let bucket = Bucket::create(&new_bucket)?;
-    /// # bucket.delete();
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn create(new_bucket: &NewBucket) -> Result<Self, Error> {
-        let url = format!("{}/b/", crate::BASE_URL);
-        let project = crate::SERVICE_ACCOUNT.project_id.clone();
-        let query = [("project", project)];
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<Self> = client
-            .post(&url)
-            .headers(crate::get_headers()?)
-            .query(&query)
-            .json(new_bucket)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Returns all `Bucket`s within this project.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    ///
-    /// let buckets = Bucket::list()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn list() -> Result<Vec<Self>, Error> {
-        let url = format!("{}/b/", crate::BASE_URL);
-        let project = crate::SERVICE_ACCOUNT.project_id.clone();
-        let query = [("project", project)];
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<ListResponse<Self>> = client
-            .get(&url)
-            .headers(crate::get_headers()?)
-            .query(&query)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s.items),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Returns a single `Bucket` by its name. If the Bucket does not exist, an error is returned.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    /// # use cloud_storage::bucket::NewBucket;
-    /// # let new_bucket = NewBucket {
-    /// #   name: "cloud-storage-rs-doc-2".to_string(),
-    /// #    ..Default::default()
-    /// # };
-    /// # let _ = Bucket::create(&new_bucket)?;
-    ///
-    /// let bucket = Bucket::read("cloud-storage-rs-doc-2")?;
-    /// # bucket.delete()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn read(name: &str) -> Result<Self, Error> {
-        let url = format!("{}/b/{}", crate::BASE_URL, name);
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<Self> = client
-            .get(&url)
-            .headers(crate::get_headers()?)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Update an existing `Bucket`. If you declare you bucket as mutable, you can edit its fields.
-    /// You can then flush your changes to Google Cloud Storage using this method.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::bucket::{Bucket, RetentionPolicy};
-    /// # use cloud_storage::bucket::NewBucket;
-    /// # let new_bucket = NewBucket {
-    /// #   name: "cloud-storage-rs-doc-3".to_string(),
-    /// #    ..Default::default()
-    /// # };
-    /// # let _ = Bucket::create(&new_bucket)?;
-    ///
-    /// let mut bucket = Bucket::read("cloud-storage-rs-doc-3")?;
-    /// bucket.retention_policy = Some(RetentionPolicy {
-    ///     retention_period: 50,
-    ///     effective_time: chrono::Utc::now() + chrono::Duration::seconds(50),
-    ///     is_locked: Some(false),
-    /// });
-    /// bucket.update()?;
-    /// # bucket.delete()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn update(&self) -> Result<Self, Error> {
-        let url = format!("{}/b/{}", crate::BASE_URL, self.name);
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<Self> = client
-            .put(&url)
-            .headers(crate::get_headers()?)
-            .json(self)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Delete an existing `Bucket`. This permanently removes a bucket from Google Cloud Storage.
-    /// An error is returned when you don't have sufficient permissions, or when the
-    /// `retention_policy` prevents you from deleting your Bucket.
-    /// ### Example
-    /// ```no_run
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    /// # use cloud_storage::bucket::NewBucket;
-    /// # let new_bucket = NewBucket {
-    /// #   name: "unnecessary-bucket".to_string(),
-    /// #    ..Default::default()
-    /// # };
-    /// # let _ = Bucket::create(&new_bucket)?;
-    ///
-    /// let bucket = Bucket::read("unnecessary-bucket")?;
-    /// bucket.delete()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn delete(self) -> Result<(), Error> {
-        let url = format!("{}/b/{}", crate::BASE_URL, self.name);
-        let client = reqwest::blocking::Client::new();
-        let response = client.delete(&url).headers(crate::get_headers()?).send()?;
-        if response.status().is_success() {
-            Ok(())
-        } else {
-            Err(Error::Google(response.json()?))
-        }
-    }
-
-    /// Returns the [IAM Policy](https://cloud.google.com/iam/docs/) for this bucket.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    /// # use cloud_storage::bucket::NewBucket;
-    /// # let new_bucket = NewBucket {
-    /// #   name: "cloud-storage-rs-doc-4".to_string(),
-    /// #    ..Default::default()
-    /// # };
-    /// # let _ = Bucket::create(&new_bucket)?;
-    ///
-    /// let bucket = Bucket::read("cloud-storage-rs-doc-4")?;
-    /// let policy = bucket.get_iam_policy()?;
-    /// # bucket.delete()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn get_iam_policy(&self) -> Result<IamPolicy, Error> {
-        let url = format!("{}/b/{}/iam", crate::BASE_URL, self.name);
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<IamPolicy> = client
-            .get(&url)
-            .headers(crate::get_headers()?)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Updates the [IAM Policy](https://cloud.google.com/iam/docs/) for this bucket.
-    /// ### Example
-    /// ```
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    /// use cloud_storage::bucket::{IamPolicy, Binding, IamRole, StandardIamRole, Entity};
-    /// # use cloud_storage::bucket::NewBucket;
-    /// # let new_bucket = NewBucket {
-    /// #   name: "cloud-storage-rs-doc-5".to_string(),
-    /// #    ..Default::default()
-    /// # };
-    /// # let _ = Bucket::create(&new_bucket)?;
-    ///
-    /// let bucket = Bucket::read("cloud-storage-rs-doc-5")?;
-    /// let iam_policy = IamPolicy {
-    ///     version: 1,
-    ///     bindings: vec![
-    ///         Binding {
-    ///             role: IamRole::Standard(StandardIamRole::ObjectViewer),
-    ///             members: vec!["allUsers".to_string()],
-    ///             condition: None,
-    ///         }
-    ///     ],
-    ///     ..Default::default()
-    /// };
-    /// let policy = bucket.set_iam_policy(&iam_policy)?;
-    /// # bucket.delete()?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn set_iam_policy(&self, iam: &IamPolicy) -> Result<IamPolicy, Error> {
-        let url = format!("{}/b/{}/iam", crate::BASE_URL, self.name);
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<IamPolicy> = client
-            .put(&url)
-            .headers(crate::get_headers()?)
-            .json(iam)
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    /// Checks whether the user provided in the service account has this permission.
-    /// ### Example
-    /// ```no_run
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use cloud_storage::Bucket;
-    ///
-    /// let bucket = Bucket::read("my-bucket")?;
-    /// bucket.test_iam_permission("storage.buckets.get")?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn test_iam_permission(&self, permission: &str) -> Result<TestIamPermission, Error> {
-        if permission == "storage.buckets.list" || permission == "storage.buckets.create" {
-            return Err(Error::new(
-                "tested permission must not be `storage.buckets.list` or `storage.buckets.create`",
-            ));
-        }
-        let url = format!("{}/b/{}/iam/testPermissions", crate::BASE_URL, self.name);
-        let client = reqwest::blocking::Client::new();
-        let result: GoogleResponse<TestIamPermission> = client
-            .get(&url)
-            .headers(crate::get_headers()?)
-            .query(&[("permissions", permission)])
-            .send()?
-            .json()?;
-        match result {
-            GoogleResponse::Success(s) => Ok(s),
-            GoogleResponse::Error(e) => Err(e.into()),
-        }
-    }
-
-    fn _lock_retention_policy() {
-        todo!()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::resources::common::Role;
-
-    #[test]
-    fn create() -> Result<(), Box<dyn std::error::Error>> {
-        dotenv::dotenv().ok();
-        let base_name = std::env::var("TEST_BUCKET")?;
-        // use a more complex bucket in this test.
-        let new_bucket = NewBucket {
-            name: format!("{}-test-create", base_name),
-            default_event_based_hold: Some(true),
-            acl: Some(vec![NewBucketAccessControl {
-                entity: Entity::AllUsers,
-                role: Role::Reader,
-            }]),
-            default_object_acl: Some(vec![NewDefaultObjectAccessControl {
-                entity: Entity::AllUsers,
-                role: Role::Reader,
-            }]),
-            iam_configuration: Some(IamConfiguration {
-                uniform_bucket_level_access: UniformBucketLevelAccess {
-                    enabled: false,
-                    locked_time: None,
-                },
-            }),
-            ..Default::default()
-        };
-        let bucket = Bucket::create(&new_bucket)?;
-        bucket.delete()?;
-        Ok(())
-    }
-
-    #[test]
-    fn list() -> Result<(), Box<dyn std::error::Error>> {
-        Bucket::list()?;
-        Ok(())
-    }
-
-    #[test]
-    fn read() -> Result<(), Box<dyn std::error::Error>> {
-        let bucket = crate::create_test_bucket("test-read");
-        let also_bucket = Bucket::read(&bucket.name)?;
-        assert_eq!(bucket, also_bucket);
-        bucket.delete()?;
-        assert!(also_bucket.delete().is_err());
-        Ok(())
-    }
-
-    #[test]
-    fn update() -> Result<(), Box<dyn std::error::Error>> {
-        let mut bucket = crate::create_test_bucket("test-update");
-        bucket.retention_policy = Some(RetentionPolicy {
-            retention_period: 50,
-            effective_time: chrono::Utc::now() + chrono::Duration::seconds(50),
-            is_locked: Some(false),
-        });
-        bucket.update()?;
-        let updated = Bucket::read(&bucket.name)?;
-        assert_eq!(updated.retention_policy.unwrap().retention_period, 50);
-        bucket.delete()?;
-        Ok(())
-    }
-
-    // used a lot throughout the other tests, but included for completeness
-    #[test]
-    fn delete() -> Result<(), Box<dyn std::error::Error>> {
-        let bucket = crate::create_test_bucket("test-delete");
-        bucket.delete()?;
-        Ok(())
-    }
-
-    #[test]
-    fn get_iam_policy() -> Result<(), Box<dyn std::error::Error>> {
-        let bucket = crate::create_test_bucket("test-get-iam-policy");
-        bucket.get_iam_policy()?;
-        bucket.delete()?;
-        Ok(())
-    }
-
-    #[test]
-    fn set_iam_policy() -> Result<(), Box<dyn std::error::Error>> {
-        // use crate::resources::iam_policy::{Binding, IamRole, StandardIamRole};
-
-        let bucket = crate::create_test_bucket("test-set-iam-policy");
-        let iam_policy = IamPolicy {
-            bindings: vec![Binding {
-                role: IamRole::Standard(StandardIamRole::ObjectViewer),
-                members: vec!["allUsers".to_string()],
-                condition: None,
-            }],
-            ..Default::default()
-        };
-        bucket.set_iam_policy(&iam_policy)?;
-        assert_eq!(
-            bucket.get_iam_policy()?.bindings,
-            iam_policy.bindings
-        );
-        bucket.delete()?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_iam_permission() -> Result<(), Box<dyn std::error::Error>> {
-        let bucket = crate::create_test_bucket("test-test-ia-permission");
-        bucket.test_iam_permission("storage.buckets.get")?;
-        bucket.delete()?;
-        Ok(())
-    }
 }
